chore(release): add automated release workflow
Adds a workflow file which can be run to create a semantic
tag and immediately afterwards, create a release on GitHub.

Any collaborator with write access should be able to
trigger the workflow.

This commit also adds another workflow file, which leaves a
comment after a PR has been merged. The comment reminds to
trigger the aforementioned release workflow.